projects/tgi-llm/ui/src/requirements.txt
python-dotenv==1.0.0
gradio==3.47.1
huggingface-hub==0.17.3
text-generation==0.6.1
